Add files via upload
V0.9.1 is a generic version of the initial beta version of this R script, requiring user inputs for the Pure API URL path and lacking some functionality. V0.9.1_Soton is a version which is specific to University of Southampton, UK users. It includes a portion of code which looks up the EPrints (ePrints Soton) URI. This version also does not require users to enter the Pure API URL path to avoid human error, this can be modified to other local users' needs, if preferred.